Rename analysis-result.ts to analysis.ts
