+++ src/model/analysis.ts
